Create LaTeX document.
Automatically save figures(png) and tables(both in md and in tex) created by python code in the notebook;
Create a latex file;
Create subfiles and place subfiles into the main tex.
So after running the code each time, the figures and tables in the tex should be updated automatically.
Good job Z!